services/voice-training-service/requirements.txt
fastapi
uvicorn
python-multipart
torch
TTS
minio
ffmpeg-python
numpy
scipy
psycopg2-binary
prometheus-client
